--- a/tempodb/encoding/vparquet/block_traceql.go
+++ b/tempodb/encoding/vparquet/block_traceql.go
@@ -96,18 +96,19 @@ func (b *backendBlock) Fetch(ctx context.Context, req traceql.FetchSpansRequest,
 		return traceql.FetchSpansResponse{}, errors.Wrap(err, "conditions invalid")
 	}
 
-	pf, _, err := b.openForSearch(ctx, opts)
+	pf, rr, err := b.openForSearch(ctx, opts)
 	if err != nil {
 		return traceql.FetchSpansResponse{}, err
 	}
 
-	iter, err := fetch(ctx, req, pf)
+	iter, err := fetch(ctx, req, pf, opts)
 	if err != nil {
 		return traceql.FetchSpansResponse{}, errors.Wrap(err, "creating fetch iter")
 	}
 
 	return traceql.FetchSpansResponse{
 		Results: iter,
+		Bytes:   func() uint64 { return rr.TotalBytesRead.Load() },
 	}, nil
 }
 
@@ -275,7 +276,7 @@ func (i *mergeSpansetIterator) Next(ctx context.Context) (*traceql.Spanset, erro
 //                                                            |
 //                                                            V
 
-func fetch(ctx context.Context, req traceql.FetchSpansRequest, pf *parquet.File) (*spansetIterator, error) {
+func fetch(ctx context.Context, req traceql.FetchSpansRequest, pf *parquet.File, opts common.SearchOptions) (*spansetIterator, error) {
 
 	// Categorize conditions into span-level or resource-level
 	var (
@@ -314,9 +315,17 @@ func fetch(ctx context.Context, req traceql.FetchSpansRequest, pf *parquet.File)
 		}
 	}
 
-	// For now we iterate all row groups in the file
-	// TODO: Add sharding params to the traceql request?
-	makeIter := makeIterFunc(ctx, pf.RowGroups(), pf)
+	rgs := pf.RowGroups()
+	if opts.TotalPages > 0 {
+		// Read UP TO TotalPages.  The sharding calculations
+		// are just estimates, so it may not line up with the
+		// actual number of pages in this file.
+		if opts.StartPage+opts.TotalPages > len(rgs) {
+			opts.TotalPages = len(rgs) - opts.StartPage
+		}
+		rgs = rgs[opts.StartPage : opts.StartPage+opts.TotalPages]
+	}
+	makeIter := makeIterFunc(ctx, rgs, pf)
 
 	// Global state
 	// Span-filtering behavior changes depending on the resource-filtering in effect,
@@ -456,11 +465,7 @@ func createSpanIterator(makeIter makeIterFn, conditions []traceql.Condition, sta
 		endFilter = parquetquery.NewIntBetweenPredicate(int64(start), math.MaxInt64)
 	}
 
-	// Static columns that are always loaded
 	var required []parquetquery.Iterator
-	required = append(required, makeIter(columnPathSpanID, nil, columnPathSpanID))
-	required = append(required, makeIter(columnPathSpanStartTime, startFilter, columnPathSpanStartTime))
-	required = append(required, makeIter(columnPathSpanEndTime, endFilter, columnPathSpanEndTime))
 
 	minCount := 0
 	if requireAtLeastOneMatch {
@@ -490,6 +495,13 @@ func createSpanIterator(makeIter makeIterFn, conditions []traceql.Condition, sta
 		required = append(required, parquetquery.NewUnionIterator(DefinitionLevelResourceSpansILSSpan, iters, nil))
 		iters = nil
 	}
+
+	// Static columns that are always loaded
+	// Since these are always present, they are put at the end and will only
+	// be read once the core conditions are met.
+	required = append(required, makeIter(columnPathSpanStartTime, startFilter, columnPathSpanStartTime))
+	required = append(required, makeIter(columnPathSpanEndTime, endFilter, columnPathSpanEndTime))
+	required = append(required, makeIter(columnPathSpanID, nil, columnPathSpanID))
 
 	// Left join here means the span id/start/end iterators + 1 are required,
 	// and all other conditions are optional. Whatever matches is returned.
@@ -557,9 +569,7 @@ func createResourceIterator(makeIter makeIterFn, spanIterator parquetquery.Itera
 		minCount,
 	}
 
-	required := []parquetquery.Iterator{
-		spanIterator,
-	}
+	var required []parquetquery.Iterator
 
 	// This is an optimization for when all of the resource conditions must be met.
 	// We simply move all iterators into the required list.
@@ -576,6 +586,10 @@ func createResourceIterator(makeIter makeIterFn, spanIterator parquetquery.Itera
 		required = append(required, parquetquery.NewUnionIterator(DefinitionLevelResourceSpans, iters, nil))
 		iters = nil
 	}
+
+	// Put span iterator last so it is only read when
+	// the resource conditions are met.
+	required = append(required, spanIterator)
 
 	// Left join here means the span iterator + 1 are required,
 	// and all other resource conditions are optional. Whatever matches